+++ src/nature/population_operators.clj
@@ -1,8 +1,7 @@
 (ns nature.population-operators
   "Functions that span or operate against entire populations"
   (:require [nature.population-presets :as pp]
-            [bigml.sampling.simple :as bss])
-  (:gen-class))
+            [bigml.sampling.simple :as bss]))
 
 (defn keep-elite
   "Find the top `number-to-keep` individuals in `population`, and increment their ages by 1"
@@ -22,11 +21,11 @@
    Then apply functions in `unary-operator-set` to the result.
    Optionally, include a map of `settings` to guide overall behavior.
    If the `:carry-over` setting is added, the elite member `n` of the prior generation will be advanced to the next generation."
-  ([population binary-operator-set unary-operator-set]
-   (let [binary-pop (repeatedly (count population) #(apply (rand-nth binary-operator-set) [(weighted-selection-of-population population 2)]))]
-     (map #((rand-nth unary-operator-set) %) binary-pop)))
-  ([population binary-operator-set unary-operator-set settings]
+  ([population population-size binary-operator-set unary-operator-set]
+   (let [binary-pop (repeatedly population-size #(apply (rand-nth binary-operator-set) [(weighted-selection-of-population population 2)]))]
+     (map #(apply (rand-nth unary-operator-set) [%]) binary-pop)))
+  ([population population-size binary-operator-set unary-operator-set settings]
   ;; Currently, :carry-over is the only optional behavior
    (if (> (:carry-over settings) 0)
-     (concat (advance-generation population binary-operator-set unary-operator-set) (keep-elite population (:carry-over settings)))
-     (advance-generation population binary-operator-set unary-operator-set))))
+     (concat (advance-generation population (- population-size (:carry-over settings)) binary-operator-set unary-operator-set) (keep-elite population (:carry-over settings)))
+     (advance-generation population population-size binary-operator-set unary-operator-set))))