+++ src/nature/population_presets.clj
@@ -16,3 +16,15 @@
   ([top] (map inc (range top)))
   ([top bottom] (range bottom (inc top)))
   ([top bottom step] (range bottom (inc top) step)))
+
+(def sum-alleles
+  (partial apply +))
+
+(def default-sequence-length
+  25)
+
+(def default-population-size
+  50)
+
+(def default-generation-count
+  50)