+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 * **Fix** -
 * **Delete** -
 
+## v0.1.1 / 2019 January 14
+
+> This release refactors the application to make the core function, evolve, stand out more distinctly.
+
+* **Add** - Usage documents
+* **Fix** - crossover no longer wipes out some genetic sequences, :carry-over no longer balloons population
+
 ## v0.0.1 / 2019 January 12
 
 > This release builds out all initial functionality